refactor store class
